Added mapping offline tool, rxtx library, serial/java bridge prototype
Wow, this commit was too large.
Three (main) things were done in this commit.

1. Audrey's Protocol branch was merged into master.
====================================================
The code paths had changed, and it wasn't functional yet, but the
directory structure fix was deemed important.

This code does not compile yet. But it will. Right, audrey? :D

2. Offline Tool - Map Generation
=================================
Project JankStreetCar is in progress, to determine the accuracy of our
GPS, and the accuracy of google's map.

3. Serial_test
===============
Added RXTX, the serial library to get java and serial to talk, and also
added the prototype.