Deletes another copy of opencv from the repository.
With gradle, it is stronly preferred to pull dependencies in at build time,
rather keeping some cut of the source code in the source control system.

From now on, it would be reasonable to consider this practice a necessary
and reasonable part of source control hygiene. This keeps the size of the
repository small, which is necessary for CI.

If there are any troubles with bringing in dependencies at build-time,
please contact @msebek (sebek.matt@gmail.com) post-haste.