Repair Build break due to folder structure
When code restructuring for gradle occurred, an extra folder was added
to the path of BuggyRos, resulting in a build break in Alice when Alice
tried to consume the binary. This commit moves this extra layer.

In addition, this commit updates the .gitignore to include gradle
temporary files, and removed mistakenly added eclipse project files.