Fix dropped messages & reverse ordering; Add test.
Last fix (dropping the lock over the user callback) incorrectly did not drain
the work queue before going back to sleep; this meant, if messages arrived too
quickly, they would never be serviced. Now, we atomically make sure the work
queue is empty and then go to sleep; else we keep draining it.

In addition, fix an issue where we service callbacks in the wrong order.
We had treated local_inbox as a stack rather than a queue. Now, it's a stack
like it should be.

In addition, add a unit test.