I think that I got ordering of which messages to show wrong
I think that I got ordering of which messages to show wrong. 
Can someone verify this.  An exception should always be visible, a warning should only be visible for warning or note, note should only be visible for notes. 

You also may want to create a new message type for Show No Errors, and Debug.  

This entire search could be made more scalable with a list that is traversed for reporting level